+++ CHANGELOG.md
@@ -10,6 +10,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 <!-- ### Added | Changed | Deprecated | Removed | Fixed | Security -->
 
+### Changed
+
+- Connect to runner node within backend instead of depending on `DNSCluster`.
+
 <!--------------------- Don't add new entries after this line --------------------->
 
 ## [0.1.0] - 2023-12-09